Moved Vidhi's changed over to Vite, and removed the components folder from my end.
