+++ frontend/src/App.jsx
@@ -6,7 +6,7 @@ import ForgotPasswordPage from './ForgotPasswordPage';
 import DashboardPage from './DashboardPage';
 import AdminDashboard from "./AdminDashboard";
 import ProtectedAdminRoute from "./ProtectedAdminRoute";
-import GrantDetailsModal from './components/GrantDetailsModal';
+import GrantDetailsModal from './GrantDetailsModal';
 
 function App() {
     return (
