+++ frontend/src/GrantDetailsModal.jsx
