+++ frontend/src/styles/DashboardPage.css
@@ -1,12 +1,14 @@
+/* Dashboard Layout */
 .dashboard-container {
     display: flex;
     height: 100vh;
+    background: #f8f9fa;
 }
 
+/* Sidebar */
 .sidebar {
     display: flex;
     flex-direction: column;
-    justify-content: space-between;  /* Pushes logout to the bottom */
     width: 250px;
     height: 100vh;
     background: linear-gradient(to bottom, #6a0dad, #8a2be2);
@@ -16,9 +18,9 @@
 }
 
 .sidebar-logo {
-    width: 100px;
+    width: 120px;
+    height: auto;
     margin-bottom: 20px;
-    display: block;
 }
 
 .sidebar-menu {
@@ -28,55 +30,101 @@
     flex-grow: 1;
     display: flex;
     flex-direction: column;
-    gap: 50px;
+    gap: 15px;
 }
 
-.sidebar-menu li:hover {
-    background-color: rgba(255, 255, 255, 0.2);
-    border-radius: 5px;
+.sidebar-menu li {
+    padding: 10px;
+    cursor: pointer;
 }
 
 .sidebar-menu li:hover {
     background: rgba(255, 255, 255, 0.2);
-}
-
-.dashboard-content {
-    flex: 1;
-    padding: 20px;
-    background: #f7f7f7;
-    overflow-y: auto;
-}
-
-.dashboard-container {
-    display: flex;
-    height: 100vh; /* Make sure it takes full screen height */
-    overflow: hidden; /* Prevent scrolling */
-}
-
-.search-bar {
-    padding: 8px;
-    width: 100%;
-    max-width: 400px;
-    margin-bottom: 20px;
-    border: 1px solid #ccc;
     border-radius: 5px;
 }
 
+/* Content Styling */
+.dashboard-content {
+    flex: 1;
+    padding: 20px;
+    background: white;
+    overflow-y: auto;
+}
+
+/* Title */
+.dashboard-title {
+    font-size: 22px;
+    font-weight: bold;
+    margin-bottom: 15px;
+}
+
+/* Summary Box - Display Inline */
+.summary-box {
+    background: #8a2be2;  /* Purple background */
+    color: white;
+    padding: 15px;
+    border-radius: 5px;
+    font-size: 16px;
+    margin-bottom: 20px;
+    display: flex;  /* Make elements align in a row */
+    justify-content: space-between; /* Space out content */
+    align-items: center;
+    width: 100%;
+    max-width: 500px; /* Optional: Adjust width */
+}
+
+
+/* Filters */
+.filter-container {
+    display: flex;
+    gap: 10px;
+    margin-bottom: 20px;
+    flex-wrap: wrap;
+}
+
+.filter-input {
+    padding: 8px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    font-size: 14px;
+}
+
+/* Table */
 .grants-table {
     width: 100%;
     border-collapse: collapse;
     background: white;
+    border-radius: 5px;
+    overflow: hidden;
 }
 
-.grants-table th, .grants-table td {
+.grants-table th {
+    background: #8a2be2;
+    color: white;
     padding: 10px;
-    border: 1px solid #ddd;
+    text-align: left;
 }
 
+.grants-table td {
+    padding: 10px;
+    border-bottom: 1px solid #ddd;
+}
 
+.grants-table tr:hover {
+    background-color: #f1f1f1;
+}
+
+/* No Results Message */
+.no-results {
+    text-align: center;
+    font-weight: bold;
+    color: #666;
+    padding: 15px;
+}
+
+/* Logout Button */
 .logout-button {
     margin-top: auto;
-    margin-bottom: 20px;
     background: #ff4b5c;
     color: white;
     border: none;
@@ -91,5 +139,3 @@
 .logout-button:hover {
     background: #d63447;
 }
-
-
